requirements.txt
aspy.yaml==1.2.0
backports.functools-lru-cache==1.5
BeautifulSoup==3.2.1
beautifulsoup4==4.7.1
cachetools==3.0.0
certifi==2018.4.16
cfgv==1.5.0
chardet==3.0.4
Click==7.0
colorama==0.3.9
configparser==3.5.0
contextlib2==0.5.5
coverage==4.5.3
disco-py==0.0.12
enum34==1.1.6
factory-boy==2.8.1
Faker==0.9.0
flake8==3.5.0
Flask==1.0.2
funcsigs==1.0.2
future==0.16.0
futures==3.2.0
gevent==1.2.2
google-api-core==1.7.0
google-auth==1.6.2
google-cloud-vision==0.35.2
googleapis-common-protos==1.5.6
greenlet==0.4.15
grpcio==1.18.0
hiro==0.5
holster==2.0.0
honcho==1.0.1
identify==1.4.0
idna==2.6
ImageHash==4.0
importlib-metadata==0.8
importlib-resources==1.0.2
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
lxml==4.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
meld3==1.0.2
mock==3.0.5
nodeenv==1.3.3
numpy==1.16.0
opencv-python==4.0.0.21
pathlib2==2.3.3
pbr==4.0.3
pefile==2018.8.8
pep8==1.7.1
Pillow==5.4.1
pre-commit==1.14.4
protobuf==3.6.1
psycopg2==2.7.7
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.3.1
pyflakes==1.6.0
pytesseract==0.2.6
pytest==3.1.2
pytest-cov==2.5.1
pytest-flask==0.10.0
pytest-mock==1.9.0
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.1
PyYAML==3.13
requests==2.18.4
rsa==4.0
scandir==1.9.0
scipy==1.2.1
six==1.11.0
soupsieve==1.8
SQLAlchemy==1.3.4
SQLAlchemy-Utils==0.33.3
sqlitedict==1.6.0
stevedore==1.28.0
supervisor==3.3.5
text-unidecode==1.2
toml==0.10.0
typing==3.6.6
urllib3==1.22
virtualenv==16.3.0
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.2
websocket-client==0.44.0
Werkzeug==0.14.1
zipp==0.3.3
